Style 04 flex
Correct solution. Their solution is by and far more elegant. I suspect
my solution is redundant in some parts and perhaps in others provides
good structure that is simply not necessary but also not bad if I wanted
to elaborate later. However, it's clear I do not currently know the
difference.